ucm-laptop-recommended/requirements.txt
spade==3.2.2
python-dotenv==0.19.2
